Solved Recursive Chamber Division
After rewriting the way information is passed down through each division cycle, the issue of rooms appearing outside of the expected area was isolated to a single operation that was returning a negative position. This bug was difficult to narrow down as it only became a problem after enough cycles happened.